+++ src/web_watchdog/email.clj
@@ -29,7 +29,7 @@
                  (:url new-site)
                  (-> new-site :title escape-html))
          (format "<p>Previous content (from %s): <span class='content'>%s</span></p>"
-                 (-> old-site :state :last-change-utc utils/epoch->now-aware-str)
+                 (-> old-site :state :last-change-time utils/epoch->now-aware-str)
                  (-> old-site :state :content-snippet escape-html))
          (format "<p>New content: <span class='content'>%s</span></p>"
                  (-> new-site :state :content-snippet escape-html))